requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
hyperspy>=1.7.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# dflow and workflow (Python 3.10 compatible version)
# Note: dflow 1.7.0+ requires Python >=3.11
# For Python 3.10, try dflow 1.6.x first, if that fails, try 1.5.x or earlier
# If dflow installation still fails, you may need to install it separately:
# pip install "dflow>=1.5.0,<1.7.0"
pydflow>=1.5.0,<1.7.0
kubernetes>=24.0.0

# Data processing
h5py>=3.7.0
mrcfile>=1.4.0
tifffile>=2022.5.4

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0
pyvista>=0.38.0

# Visualization (optional - 3D rendering)
# Note: mayavi requires C++ compiler on Windows (Microsoft Visual C++ 14.0+)
# If you don't need 3D isosurface rendering, you can skip mayavi
# Install it separately if needed: pip install mayavi
# mayavi>=4.7.4  # Commented out - install separately if needed

# Monitoring and logging
prometheus-client>=0.15.0
python-json-logger>=2.0.0

# CLI
click>=8.1.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
mypy>=0.991

# Code quality
black>=22.10.0
flake8>=5.0.0
isort>=5.11.0

# Documentation
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.0

# Utilities
psutil>=5.9.0
tqdm>=4.64.0

